backend/requirements.txt
# =============================================================================
# requirements.txt - Backend Python Bağımlılıkları
# Octoqus - Harcama Takip & Sosyal Finans Platformu
# Tüm versiyonlar sabitlenmiştir (production güvenliği)
# =============================================================================

# --- Web Framework ---
sanic==23.12.1                  # Async HTTP + WebSocket framework
sanic-ext==23.12.0              # Dependency injection, OpenAPI desteği

# --- ORM & Database ---
sqlalchemy[asyncio]==2.0.30     # Async ORM (asyncio engine)
aiomysql==0.2.0                 # SQLAlchemy async MySQL driver
alembic==1.13.1                 # Veritabanı migration yönetimi

# --- Redis ---
redis[hiredis]==5.0.4           # Async Redis client (cache, pub/sub, rate limit)

# --- Auth & Security ---
pyjwt==2.8.0                    # JSON Web Token
bcrypt==4.1.3                   # Şifre hashing

# --- Validation & Serialization ---
pydantic==2.7.1                 # Data validation & schema
email-validator==2.1.1          # Pydantic EmailStr desteği için

# --- File & Image Processing ---
aiofiles==23.2.1                # Async dosya okuma (upload pipeline)
Pillow==10.3.0                  # Image processing and optimization

# --- Export ---
xlsxwriter==3.2.0               # Excel file generation
fpdf2==2.7.9                    # PDF generation

# --- Email ---
aiosmtplib==3.0.1               # Async SMTP client (Şifre sıfırlama vb.)

# --- Utilities ---
python-dotenv==1.0.1            # .env yükleme (geliştirme ortamı)
ujson==5.9.0                    # Ultra-fast JSON serialization
python-multipart==0.0.9         # Multipart form / file upload desteği

# --- Logging ---
structlog==24.1.0               # Yapılandırılmış loglama (JSON output)

# --- Testing (CI/CD) ---
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-sanic==1.9.1
httpx==0.27.0                   # Test HTTP client
factory-boy==3.3.0              # Test fixture factory
